Move problematic unit test to newton tests
